pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
requires-python = ">=3.10, <3.15"
name = "projet_etudes"
readme = "README.md"
dynamic = ["version"]
dependencies = [
    "ipython>=8.10",
    "jupyterlab>=3.0",
    "notebook",
    "kedro[jupyter]~=1.0.0",
    "kedro-datasets[pandas-csvdataset, pandas-exceldataset]",
    "kedro-viz>=6.7.0",
    "scikit-learn~=1.5.1",
    "seaborn~=0.12.1",
    "setuptools; python_version >= '3.12'",
    "pymongo",
    "python-dotenv",
    "pandas",
    "atproto==0.0.63",
    "kedro-airflow",
    "fastapi",
    "pydantic",
    "mysql-connector-python",
    "nltk",
    "pyspark",
    "google-genai",
    "claude-agent-sdk",
    "streamlit",
    "prompt-toolkit==3.0.51",
    "prometheus-fastapi-instrumentator",
    "tensorflow",
    "codecarbon",
    "commitizen>=4.8.3",
]

[project.scripts]
"projet-etudes" = "projet_etudes.__main__:main"

[project.entry-points."kedro.hooks"]

[project.optional-dependencies]
docs = [
    "docutils<0.21",
    "sphinx>=5.3,<7.3",
    "sphinx_rtd_theme==2.0.0",
    "nbsphinx==0.8.1",
    "sphinx-autodoc-typehints==1.20.2",
    "sphinx_copybutton==0.5.2",
    "ipykernel>=5.3, <7.0",
    "Jinja2<3.2.0",
    "myst-parser>=1.0,<2.1"
]
dev = [
    "pytest-cov~=3.0",
    "pytest-mock>=1.7.1, <2.0",
    "pytest~=7.2",
    "ruff~=0.12.0",
    "black",
    "pre-commit",
    "commitizen"
]

[tool.setuptools.dynamic]
version = {attr = "projet_etudes.__version__"}

[tool.setuptools.packages.find]
where = ["src"]
namespaces = false

[tool.kedro]
package_name = "projet_etudes"
project_name = "projet_etudes"
kedro_init_version = "1.0.0"
tools = "['Linting', 'Testing', 'Custom Logging', 'Documentation', 'Data Structure', 'PySpark']"
example_pipeline = "False"
source_dir = "src"

[tool.pytest.ini_options]
addopts = """
--cov-report term-missing \
--cov src/projet_etudes -ra"""

[tool.coverage.report]
fail_under = 0
show_missing = true
exclude_lines = ["pragma: no cover", "raise NotImplementedError"]

[tool.black]
target-version = ["py310"]
line-length = 88

[tool.ruff.format]
docstring-code-format = true

[tool.ruff]
line-length = 88
show-fixes = true

[tool.ruff.lint]
select = [
    "F",   # Pyflakes
    "W",   # pycodestyle
    "E",   # pycodestyle
    "I",   # isort
    "UP",  # pyupgrade
    "PL",  # Pylint
    "T201", # Print Statement
]
ignore = [
    "E501",     # line-too-long — handled by black
    "PLR0913",  # too-many-arguments — common in framework callbacks
    "PLW0603",  # global-statement — intentional lazy singleton pattern
]

[tool.kedro_telemetry]
project_id = "b02942fb0e324d578b73d8185cd058d2"
